Move descriptor::pipeline_layout to pipeline::layout
