+++ vulkano/src/command_buffer/auto.rs
@@ -38,7 +38,6 @@ use crate::command_buffer::StateCacherOutcome;
 use crate::command_buffer::SubpassContents;
 use crate::descriptor::descriptor::{DescriptorBufferDesc, DescriptorDescTy};
 use crate::descriptor::descriptor_set::{DescriptorSetDesc, DescriptorSetsCollection};
-use crate::descriptor::pipeline_layout::PipelineLayout;
 use crate::device::Device;
 use crate::device::DeviceOwned;
 use crate::device::Queue;
@@ -51,6 +50,7 @@ use crate::image::ImageAspects;
 use crate::image::ImageLayout;
 use crate::instance::QueueFamily;
 use crate::pipeline::input_assembly::Index;
+use crate::pipeline::layout::PipelineLayout;
 use crate::pipeline::vertex::VertexSource;
 use crate::pipeline::ComputePipelineAbstract;
 use crate::pipeline::GraphicsPipelineAbstract;
